pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regulatory-knowledge-base"
version = "0.1.0"
description = "Graph knowledge base for regulatory guidance documents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Regulatory KB Team"}
]
dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "PyPDF2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "spacy>=3.7.0",
    "redis>=5.0.0",
    "falkordb>=1.0.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.92.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
]
cdk = [
    "aws-cdk-lib>=2.120.0",
    "constructs>=10.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
